Fix for a problem that was revealed/introduced when support for multiple images was added to TB-Loader.
The TB-Loader logs the images (content packages) that are deployed to a
Talking Book. When multiple images are deployed at once (onto a single Talking
Book), they are all (two on TBv1) listed, separated by a comma. That causes
csv parsing to shift subsequent fields. The fix is to quote the list of
images, in the TB-Loader.

This update consists of a hack to detect and work around the bug.